Versión 2.0

+++ src/main/java/com/gasto/gasto/Controlador/GestorController.java
@@ -3,12 +3,18 @@ package com.gasto.gasto.Controlador;
 import com.gasto.gasto.Excepciones.RequestException;
 import com.gasto.gasto.Modelo.Gestor;
 import com.gasto.gasto.Service.GestorService;
+import com.gasto.gasto.utils.JWTUtil;
+import de.mkammerer.argon2.Argon2;
+import de.mkammerer.argon2.Argon2Factory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.lang.Nullable;
+import org.springframework.web.bind.MissingRequestHeaderException;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.constraints.NotNull;
+import java.util.Date;
 import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
@@ -33,6 +39,9 @@ public class GestorController {
     @Autowired
     private GestorService gestorService;
 
+    @Autowired
+    private AuthController authController;
+
     /**
      * Constructor que recibe un servicio de Gestor.
      *
@@ -50,7 +59,10 @@ public class GestorController {
      * @throws RequestException si no se encuentran gestores en la base de datos.
      */
     @GetMapping
-    public ResponseEntity<List<Gestor>> findAll() {
+    public ResponseEntity<List<Gestor>> findAll(@RequestHeader(value = "Authorization") @Nullable String token) {
+        if(!authController.validarSesion(token)){
+            throw new RequestException("", HttpStatus.FORBIDDEN, "Token no valido");
+        }
         if (gestorService.getAll().isEmpty()) {
             throw new RequestException("G-106A", HttpStatus.NOT_FOUND, "No se encontraron gestores en la base de datos");
         }
@@ -67,7 +79,10 @@ public class GestorController {
      * @throws RequestException si no se encuentra un gestor con el ID proporcionado.
      */
     @GetMapping("/{id}")
-    public ResponseEntity<Gestor> findById(@PathVariable Long id) {
+    public ResponseEntity<Gestor> findById(@RequestHeader(value = "Authorization") @Nullable String token, @PathVariable Long id) {
+        if(!authController.validarSesion(token)){
+            throw new RequestException("", HttpStatus.FORBIDDEN, "Token no valido");
+        }
         Optional<Gestor> gestor = gestorService.findById(id);
         if (gestor.isPresent()) {
             return new ResponseEntity<>(gestor.get(), HttpStatus.OK);
@@ -87,9 +102,18 @@ public class GestorController {
      * si el email no es válido o está vacío, o si el rol no es Administrador o Gestor.
      */
     @PostMapping
-    public ResponseEntity<Gestor> save(@RequestBody @NotNull Gestor gestor) {
+    public ResponseEntity<Gestor> save(@RequestHeader(value = "Authorization") @Nullable String token, @RequestBody @NotNull Gestor gestor) {
+        if(!authController.validarSesion(token)){
+            throw new RequestException("", HttpStatus.FORBIDDEN, "Token no valido");
+        }
+        if (gestor.getId() == null){
+            throw new RequestException("1", HttpStatus.BAD_REQUEST,"El id no puede estar vacio");
+        }
         if(gestorService.findById(gestor.getId()).isPresent()){
             throw new RequestException("G-101",HttpStatus.BAD_REQUEST,"El id de gestor ya se encuentra en la base");
+        }
+        if (gestor.getPassword()=="" || gestor.getPassword()==null){
+            throw new RequestException("2", HttpStatus.BAD_REQUEST, "La contraseña no puede ser vacia o nula");
         }
         if(gestor.getNombre().isEmpty()){
             throw new RequestException("G-104",HttpStatus.BAD_REQUEST,"El nombre no puede estar vacio");
@@ -97,12 +121,20 @@ public class GestorController {
         if (gestor.getEmail() == null || gestor.getEmail().equals("")){
             throw new RequestException("G-102A",HttpStatus.BAD_REQUEST,"El e-mail ingresado no es valido o esta vacio");
         }
+        if (gestorService.getGestorByEmail(gestor.getEmail()).isPresent()){
+            throw new RequestException("3",HttpStatus.BAD_REQUEST,"El e-mail ya esta registrado en la base de datos");
+        }
         if (!gestor.getEmail().matches("\\b[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\\.[A-Z|a-z]{2,}\\b")) {
             throw new RequestException("G-102B", HttpStatus.BAD_REQUEST, "El correo electrónico ingresado no es válido");
         }
         if (!Objects.equals(gestor.getRol(), "Administrador") && !Objects.equals(gestor.getRol(), "Gestor")){
             throw new RequestException("G-103",HttpStatus.BAD_REQUEST, "El rol solo puede ser Administrador o Gestor");
         }
+
+        Argon2 argon2 = Argon2Factory.create(Argon2Factory.Argon2Types.ARGON2id);
+        String hash = argon2.hash(1,1024,1, gestor.getPassword());
+        gestor.setPassword(hash);
+        gestorService.save(gestor);
         return ResponseEntity.ok().build();
     }
 
@@ -116,11 +148,24 @@ public class GestorController {
      * si el email no es válido o está vacío, o si el rol no es Administrador o Gestor.
      */
     @PutMapping("/{id}")
-    public ResponseEntity<Gestor> update(@PathVariable Long id, @RequestBody Gestor gestor) {
+    public ResponseEntity<Gestor> update(@RequestHeader(value = "Authorization") @Nullable String token, @PathVariable Long id, @RequestBody Gestor gestor) {
+        if(!authController.validarSesion(token)){
+            throw new RequestException("", HttpStatus.FORBIDDEN, "Token no valido");
+        }
+        if (id == null){
+            throw new RequestException("1", HttpStatus.BAD_REQUEST,"El id no puede estar vacio");
+        }
+        if (id==99999 && !Objects.equals(gestor.getEmail(), "admin@gasto.com")){
+            throw new RequestException("",HttpStatus.BAD_REQUEST,"Este es un administrador no modificable");
+        }
         Optional<Gestor> currentGestor = gestorService.findById(id);
+        Optional<Gestor> currentGestorEmail = gestorService.getGestorByEmail(gestor.getEmail());
         if (currentGestor.isPresent()) {
             if(gestor.getNombre().isEmpty()){
                 throw new RequestException("G-104",HttpStatus.BAD_REQUEST,"El nombre no puede estar vacio");
+            }
+            if (gestor.getPassword()=="" || gestor.getPassword()==null){
+                throw new RequestException("2", HttpStatus.BAD_REQUEST, "La contraseña no puede ser vacia o nula");
             }
             if (gestor.getEmail() == null || gestor.getEmail().equals("")){
                 throw new RequestException("G-102A",HttpStatus.BAD_REQUEST,"El e-mail ingresado no es valido o esta vacio");
@@ -128,13 +173,19 @@ public class GestorController {
             if (!gestor.getEmail().matches("\\b[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\\.[A-Z|a-z]{2,}\\b")) {
                 throw new RequestException("G-102B", HttpStatus.BAD_REQUEST, "El correo electrónico ingresado no es válido");
             }
+            if (currentGestorEmail.isPresent() && !Objects.equals(currentGestorEmail.get().getId(), id)){
+                throw new RequestException("3",HttpStatus.BAD_REQUEST,"El e-mail ya esta registrado en la base de datos");
+            }
             if (!Objects.equals(gestor.getRol(), "Administrador") && !Objects.equals(gestor.getRol(), "Gestor")){
                 throw new RequestException("G-103",HttpStatus.BAD_REQUEST, "El rol solo puede ser Administrador o Gestor");
             }
             Gestor updatedGestor = currentGestor.get();
             updatedGestor.setNombre(gestor.getNombre());
             updatedGestor.setEmail(gestor.getEmail());
-            updatedGestor.setPassword(gestor.getPassword());
+            updatedGestor.setRol(gestor.getRol());
+            Argon2 argon2 = Argon2Factory.create(Argon2Factory.Argon2Types.ARGON2id);
+            String hash = argon2.hash(1,1024,1, gestor.getPassword());
+            updatedGestor.setPassword(hash);
             gestorService.save(updatedGestor);
             return new ResponseEntity<>(updatedGestor, HttpStatus.OK);
         } else {
@@ -151,8 +202,14 @@ public class GestorController {
      * @throws RequestException si no se encuentra un Gestor con el ID proporcionado.
      */
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteById(@PathVariable Long id) {
+    public ResponseEntity<Void> deleteById(@RequestHeader(value = "Authorization") @Nullable String token,@PathVariable Long id) {
+        if(!authController.validarSesion(token)){
+            throw new RequestException("", HttpStatus.FORBIDDEN, "Token no valido");
+        }
         Optional<Gestor> gestor = gestorService.findById(id);
+        if (id==99999){
+            throw new RequestException("",HttpStatus.BAD_REQUEST,"No se puede eliminar a este administrador");
+        }
         if (gestor.isPresent()) {
             gestorService.deleteById(id);
             return new ResponseEntity<>(HttpStatus.NO_CONTENT);
@@ -161,3 +218,4 @@ public class GestorController {
         }
     }
 }
+
